requirements.txt
keras_vggface==0.6
scipy==1.3.1
mtcnn==0.1.0
face_recognition==1.2.3
numpy==1.17.3
pandas==0.25.3
pytesseract==0.3.1
Werkzeug==0.16.0
textdistance==4.1.5
tensorflow==1.14.0
opencv_contrib_python==4.1.2.30
pdf2image==1.10.0
matplotlib==3.1.1
Keras==2.3.1
Flask==1.1.1
imutils==0.5.3
Pillow==7.0.0
